modules/lcr, modules_k/regex: Fix stack overflow from pcre_fullinfo(PCRE_INFO_SIZE)
Fix instances where pcre_fullinfo(PCRE_INFO_SIZE) was given only an "int"
argument, which results in stack overflow on some 64-bit platforms where
an "int" is only 32 bits long. It expects a "size_t" argument, which is
64 bits long.

+++ modules_k/regex/regex_mod.c
@@ -260,7 +260,7 @@ static int load_pcres(int action)
 	char line[FILE_MAX_LINE];
 	char **patterns = NULL;
 	pcre *pcre_tmp = NULL;
-	int pcre_size;
+	size_t pcre_size;
 	int pcre_rc;
 	const char *pcre_error;
 	int pcre_erroffset;
